+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheetHeader.js
@@ -26,6 +26,7 @@ export const DataSpreadsheetHeader = forwardRef(
       activeCellCoordinates,
       cellSize,
       columns,
+      currentMatcher,
       defaultColumn,
       headerGroups,
       scrollBarSize,
@@ -39,6 +40,7 @@ export const DataSpreadsheetHeader = forwardRef(
       updateActiveCellCoordinates,
       setHeaderCellHoldActive,
       headerCellHoldActive,
+      visibleColumns,
     },
     ref
   ) => {
@@ -61,6 +63,7 @@ export const DataSpreadsheetHeader = forwardRef(
     const handleColumnHeaderClick = (index) => {
       return (event) => {
         const isHoldingCommandKey = event.metaKey || event.ctrlKey;
+        const isHoldingShiftKey = event.shiftKey;
         handleHeaderCellSelection({
           type: 'column',
           activeCellCoordinates,
@@ -73,6 +76,8 @@ export const DataSpreadsheetHeader = forwardRef(
           index,
           setSelectionAreaData,
           isHoldingCommandKey,
+          isHoldingShiftKey,
+          currentMatcher,
         });
       };
     };
@@ -91,19 +96,35 @@ export const DataSpreadsheetHeader = forwardRef(
 
     const handleHeaderMouseDown = (index) => {
       return (event) => {
+        if (event.shiftKey) {
+          // Remove columns, need to call handleHeaderCellSelection
+          return;
+        }
         setSelectedHeaderReorderActive(true);
-        const clickXPosition = event.clientX;
-        const headerButtonCoords = event.target.getBoundingClientRect();
-        const offsetXValue = clickXPosition - headerButtonCoords.left;
-        const bodyContainer = document.querySelector(
-          `.${blockClass}__list--container`
-        ).firstElementChild;
         const selectionAreaToClone = selectionAreas.filter(
-          (item) => item?.header?.index === index
+          (item) => item?.matcher === currentMatcher
         );
         const selectionAreaElement = ref.current.querySelector(
           `[data-matcher-id="${selectionAreaToClone[0]?.matcher}"]`
         );
+        const clickXPosition = event.clientX;
+        const headerButtonCoords = event.target.getBoundingClientRect();
+        const headerIndex = event.target.getAttribute('data-column-index');
+        const offsetXValue = clickXPosition - headerButtonCoords.left;
+        const lowestColumnIndexFromSelectionArea = Math.min(
+          selectionAreaToClone[0].point1.column,
+          selectionAreaToClone[0].point2.column
+        );
+        const selectionAreaCoords =
+          selectionAreaElement.getBoundingClientRect();
+        const updatedOffsetDifference =
+          lowestColumnIndexFromSelectionArea < parseInt(headerIndex)
+            ? offsetXValue +
+              (headerButtonCoords.left - selectionAreaCoords.left)
+            : offsetXValue;
+        const bodyContainer = document.querySelector(
+          `.${blockClass}__list--container`
+        ).firstElementChild;
         const selectionAreaClonedElement = selectionAreaElement.cloneNode();
         const reorderIndicatorLine = selectionAreaElement.cloneNode();
         reorderIndicatorLine.className = `${blockClass}__reorder-indicator-line`;
@@ -113,7 +134,7 @@ export const DataSpreadsheetHeader = forwardRef(
         );
         selectionAreaClonedElement.setAttribute(
           'data-clone-offset-x',
-          offsetXValue
+          updatedOffsetDifference
         );
         selectionAreaClonedElement.setAttribute(
           'data-column-index-original',
@@ -141,6 +162,7 @@ export const DataSpreadsheetHeader = forwardRef(
                 scrollBarSizeValue,
                 totalVisibleColumns,
                 defaultColumn,
+                visibleColumns,
               }),
               overflow: 'hidden',
             }}
@@ -181,7 +203,8 @@ export const DataSpreadsheetHeader = forwardRef(
               const selectedHeader = checkSelectedHeaderCell(
                 index,
                 selectionAreas,
-                'column'
+                'column',
+                rows
               );
               return (
                 <div
@@ -252,12 +275,17 @@ DataSpreadsheetHeader.propTypes = {
   /**
    * Specifies the cell height
    */
-  cellSize: PropTypes.oneOf(['compact', 'standard', 'medium', 'large']),
+  cellSize: PropTypes.oneOf(['xs', 'sm', 'md', 'lg']),
 
   /**
    * All of the spreadsheet columns
    */
   columns: PropTypes.array,
+
+  /**
+   * uuid that corresponds to the current selection area
+   */
+  currentMatcher: PropTypes.string,
 
   /**
    * Default spreadsheet sizing values
@@ -328,4 +356,9 @@ DataSpreadsheetHeader.propTypes = {
    * Function used to update the active cell coordinates
    */
   updateActiveCellCoordinates: PropTypes.func,
+
+  /**
+   * Array of visible columns provided by react-table useTable hook
+   */
+  visibleColumns: PropTypes.array,
 };