+++ packages/security/src/components/Panel/Panel-story.js
@@ -25,6 +25,7 @@ import {
 } from '../..';
 
 import { InlineNotification, NotificationActionButton } from '../../';
+import { Grid } from 'carbon-components-react';
 
 const closeButtonLabel = 'Close';
 
@@ -55,15 +56,17 @@ storiesOf(patterns('Panel#legacy'), module)
             rel="noopener noreferrer"
             target="_blank"
           >
-            More info
+            View replacement
           </NotificationActionButton>
         }
         kind="info"
-        subtitle="This component is now legacy. Please click to see the go-forward component"
+        subtitle="Pattern no longer supported. The pattern will remain available, but plan to migrate to the pattern replacement."
         title=""
         hideCloseButton
       />
-      <Story />
+      <Grid>
+        <Story />
+      </Grid>
     </>
   ))
   .add(